Quantify and Annotate Short Reads in R

This package provides a pipeline for the quantification and analysis of short reads.

Details

See packageDescription('QuasR') for package details.

See also

Author

Anita Lerch, Charlotte Soneson, Dimos Gaidatzis and Michael Stadler

Examples

if (FALSE) {
# see qCount, qMeth and qProfile manual pages for examples
example(qCount)
example(qMeth)
example(qProfile)
}